Выполнены рекомендации
