+++ iOS-FakeNFT-Extended/UI/Profiles/Profile/Views/ProfileLoadingView.swift
@@ -1,0 +1,24 @@
+//
+//  ProfileLoadingView.swift
+//  iOS-FakeNFT-Extended
+//
+//  Created by Дмитрий Андрианов on 13.04.2026.
+//
+
+import SwiftUI
+
+struct ProfileLoadingView: View {
+    var body: some View {
+        VStack {
+            Spacer()
+            ProgressView()
+            Spacer()
+        }
+        .frame(maxWidth: .infinity, maxHeight: .infinity)
+        .background(Color.dayNightWhite.ignoresSafeArea())
+    }
+}
+
+#Preview {
+    ProfileLoadingView()
+}
